Morph DOM nodes when restoring stream items (#3111)
This is a followup to #3070 and #3035.

When adjusting the stream component restore functionality in #3070, I
accidentally broke live components in streams, but the tests did not
detect this because they only checked the id of the stream items and not
their content. Therefore, this commit also adjusts the tests.

Another problem that was introduced in #3035 by not leaving re-inserted
items in the DOM is that nested streams inside live components inside
streams would not be restored properly. This commit addresses this by
recursively running morphdom on the restored stream items.