Fix magic ID optimization skipping static roots. Closes #3117
This only happened for static roots without dynamic content,
inside a live component, rendered within a for comprehension.
We were not considering newly added components (ones we have
not seen before) as reset the first time we see them, and
would skip them even before they existed prior in the DOM.
The fix is to consider components as reset the first time we
see them when merging diffs, such as on new joins.