Assign index when nesting inside a list occurs.
This allows the use of nesting inside a list on a select multiple field by inserting the index.

Why: Plug does not support nesting without index, see: https://github.com/elixir-plug/plug/issues/1210
It expects either foo[] or foo[0][bar]

This is fine if you have multiple input fields, you can assign a name with an index. However, on a select multiple, you are not able to insert a different name for each option.

Automatically adding the indexing when nesting is detected allows this.