fix flaky upload redirect test
In render_upload we're using a monitor to check if a upload callback
redirects. The problem was that the message could could not have arrived
yet when we tried to receive it, therefore calling render, which would
fail as the process was gone. The fix is to always call render and check
for the DOWN message when we catch the exit.