Fix flaky unit tests (#3165)
* fix flaky upload redirect test

In render_upload we're using a monitor to check if a upload callback
redirects. The problem was that the message could could not have arrived
yet when we tried to receive it, therefore calling render, which would
fail as the process was gone. The fix is to always call render and check
for the DOWN message when we catch the exit.

* make tests async again

https://github.com/phoenixframework/phoenix_live_view/commit/de1a2cd037314277e416ba59ae7b39c675d1aa59
changed lots of tests to run synchronously, but the reason for this was
removed later in https://github.com/phoenixframework/phoenix_live_view/commit/8f9d5605a9ca7de4a4a8629c8ba08da4c10e4267.

Therefore, this commit reverts the tests to run asynchronously again.
Speeeeeed!

* do not assume the redirect already happened

* fix flaky start_async / assign_async monitor tests

It could happen that we tried to monitor the async process before it was
started, therefore getting :noproc messages. This commit adjusts the
test process to wait for a message from the async process before starting
to monitor it.