Fix bug in client file upload guide
The prose says we are changing the `presign_url` function, but really it
means to to `presign_upload`.  This becomes evident when reading through
the whole context, but can be very confusing if you don't.  I suspect
the function was originally called `presign_url` but got refactored to
`presign_upload` and the prose got missed.