Allow partially resetting an async result when using multiple assign
In circumstances where `assign_async/4` is being used to async assign
multiple results, it is possible that some of the results should persist
during loading, and others should not.

In this case, a list of keys can be passed to the `reset` option, and
only those keys will be reset during the async operation.