Support custom reloader in Phoenix.LiveView.Channel
In [Phoenix Playground](https://github.com/phoenix-playground/phoenix_playground)
I'm using a custom code reloader and I'd like to be able to use when
doing state-preserving reloads.

(I'm using a custom code reloader because the stock one,
Phoenix.CodeReloader, needs to be running in a regular Mix project.)

I believe this is pretty niche feature so I'm not sure if it's worth
documenting and if so where exactly.

I went ahead and rewrote the integration test and made it self-contained
(I defined the minimal endpoint inline) since I believe that's easier to
follow.