Cleanup collect_submitter, pass in fields
Added `select_node_from_list` because select_node
looked in the children of the passed in node.

Extracted `select_node_by_text`
Extracted `filtered_fields`

Tried to make it clearer that named inputs are
used separately.

And removed the form name/id comparison
from before which was a bit hacky.