Fix order of ack on LC nav and prevent phx-remove event propagation.
Closes #3306

While we use refs (vector clocks) and ack's to
prevent races like double form submissions and
duplicate interactions for latent clients, there
existed two bugs – one on the server, and one on
client, that could allow for an ack to be processed
too early in some conditions involving LiveComponents
with live navigation.

On the server, a LiveComponent would issue a nav like
push_patch, and we we would ack the LiveComponent
event before the parent's handle_params diff event.
This caused a race where the LiveComponent is ack'd
and pending states undone too early on the client,
before the parent's handle_params diff has been sent.

This allows scenarios where the diff from the parent
LiveView handle_params removes the LiveComponent
from the page, but there is a window where the LC is
interactive since we ack'd it early. The fix on the
server is to still send the LiveComponent diff
on redirect, but send it out of band, and ack the
original LiveComponent event within the parent's
handle_params diff that will necessarily be triggered.

On the client, a bug existed for phx-remove.
Elements are delayed from being removed from the
DOM to allow a short transition to occur, such as
a fadeout of 200-300ms. Within this window, the
elements remained interactive, which allowed for
an "unlocked" element like a form button to be
clicked after it was processed and ack'd by the server
during the phx-remove transition. The fix for the
client race is to prevent event propagation on the
removed elements during the phx-remove transition.