Fix form recovery when targeting LiveComponents
In #3201 we changed the way form recovery works by using a document
fragment to dispatch the form recovery events. When traversing the DOM
to check if a LC is in the same LiveView, this broke the check, because
when iterating over node.parentNode, we would never arrive at the
fragment itself. This commit fixes this by checking by using the
framgnet's first child instead, which is the actual LV element.

Fixes #3314.