do not emit change event while inside composition
Safari breaks character composition mode if an input is patched while
a composition is in progress. This commit works around this by ignoring
input events that are triggered while a composition is in progress and
emitting the event only after the composition is finished.

Fixes https://github.com/phoenixframework/phoenix_live_view/issues/3322.