Fix unnecessary remount when navigating back
Initially reported here: https://elixirforum.com/t/navigating-using-patch-causes-liveview-to-re-mount-when-using-browser-back-button/62595

In the popstate event LV compares the id in the history with the id of
the main view id. But on the initial navigation, the history state is
never set, so navigating all the way back causes an unnecessary remount.

This is fixed by using replaceRootHistory when joining the main view.