Keep all old values in `__changed__` when using `assign/3` instead of only maps (#3392)
This commit is a prerequisite to perform nested change tracking of other types of data than maps, eg. lists or tuples. Previously `socket.assigns.__changed__` kept old versions of maps for nested change tracking and `true` for all other types of props. Worth to note that `changed?/2` is simply checking if key is in the `__changed__` map, value itself is not used.