Add compile time warnings for attr arity mismatch
This commit introduces compile time warnings for component attributes
that are defined using `{:fun, arity}`. In the cases that we can detect
the arity at compile time, we will warn the user if the arity of the
passed function does not match the arity declared for the attribute.

Relates to #3355.