List component systems on README
A component system is often on the top of the list of required LiveView features, so let's list them in the README.

We can add new systems as well. If any system is no longer developer, we can remove it. I probably want to duplicate this list in Phoenix docs later on.

/cc @chrisgreg @woylie @nhobes @ArthurClemens @mitkins @bluzky