Add doc section on accessing a nested field's value
When using `<.inputs_for>`, accessing the raw value of the nested field
and doing calculations on it has some pitfalls, because the value can
be changesets, data or even raw parameters, depending on what is currently
happening in the form. This leads to complex code that can be avoided by
doing such calculations on the changeset instead. This commit adds a
warning adomonition block mentioning this together with a small example.

Closes #3461.