Refactor error handling.
For plug_status errors during connected mount, we now sign an instruction
within an cookie to instruct the dead render to trigger the plug error pipeline.

In dev, this will show the plug debugger with stacktrace. In prod, this will
show the 400..50 error pages.

For child LV's, we now attempt to remount a failed child up to a configurable
maximum, then failsafe keep the child in a dead state.

Co-authored-by: Steffen Deusch <steffen@deusch.me>