Reserve `{` outside of attributes
Before this patch, this was valid:

    <div id={id}>{content}</div>

and the text `{content}` is rendered _as is_, without
interpolation. I'd like to reserve this syntax for later
and for now raise an error:

    iex> import Phoenix.Component, only: [sigil_H: 2]
    iex> assigns = %{}; ~H"<div id={id}>{content}</div>"
    ** (Phoenix.LiveView.Tokenizer.ParseError) iex:3:14: `{` outside of attributes is reserved, use `\\{` to escape
      |
    1 | <div id={id}>{content}</div>
        (phoenix_live_view 1.0.0-rc.7) lib/phoenix_live_view/tokenizer.ex:732: Phoenix.LiveView.Tokenizer.raise_syntax_error!/3
        (phoenix_live_view 1.0.0-rc.7) lib/phoenix_live_view/tag_engine.ex:295: Phoenix.LiveView.TagEngine.handle_text/3
        (eex 1.18.0-dev) lib/eex/compiler.ex:331: EEx.Compiler.generate_buffer/4
        (phoenix_live_view 1.0.0-rc.7) expanding macro: Phoenix.Component.sigil_H/2
        iex:2: (file)

If someone needs to use `{`, there's a way to escape it using `\\{`:

    iex> assigns = %{}; ~H"<div id={:foo}>\\{content}</div>".static
    ["<div", ">\\{content}</div>"]

The idea is to eventually use the same interpolation syntax across the
board. And when that happens, deprecate `<%= %>` in favour of `{ }`.
Supporting new syntax as well as deprecating old syntax is a massive
change because of existing Phoenix projects, syntax highlighters, etc.
I think we should be able to automate moving from `<% %>` to `{ }`
using tools like https://github.com/adobe/elixir-styler.