clear private HOOK_ID as soon as hook is destroyed
Fixes #3496.

When navigating from a sticky LiveView to another page that contains
a hook on an element with the same id, morphdom would merge the old
element into the new one and still find the old HOOK_ID in the element's
DOM.private data, leading to the new hook not initializing properly.