always merge defaultArgs in JS.exec (#3499)
Fixes #3494.

I don't know why, but when passing "defaults" to JS.exec, we only merged
the defaults if there is a "data" property. If someone used

    <form phx-submit={JS.push("save")}>
      <button name="foo" value="baz">Submit</button>
    </form>

the submitter would be passed as defaults, but the command itself would
be `["push", {"event": "save"}]`, therefore the defaults would not be
merged. This commit changes the merging to always happen, but the
phxEvent args always override the defaults.