properly clear viewHooks when destroying a hook (#3628)
Relates to: #3496
Fixes: #3623

In #3496, I cleared the hookId on the element in destroyed,
but the view's destroyHook function actually relied on the hookId to
clean up the view's viewHooks object. This would cause hooks' destroy
function to be called multiple times, as the view was not able to remove
old hooks.