fix race conditions when upload progress redirects (#3676)
When someone does a push_navigate in the progress callback of an upload,
the render_upload function of LiveViewTest sometimes did not properly
return the live_redirect error tuple. This happened because the Upload-
Channel closes as soon as an upload is consumed. When this happens in the
progress callback, the test code and the channel code raced:
The channel sends a redirect message to itself (or the root view) which
shuts down once it receives it. The test code triggers a new render
through the ClientProxy. Now it could happen that:

1. the test process can successfully call the ClientProxy, but the
   ClientProxy then crashed while pinging the channel
2. or the ClientProxy succeeds in pinging and actually returns the new
   HTML while the test expects the redirect error tuple
3. or the channel and client proxy shut down in time for the test process
   to fail rendering and return the error tuple as expected

We fix this race by ensuring that we ping the view (and root view) as
soon as the upload is finished. This ensures that the channel has time to
send the redirect message before the render call can ping it again. Thus,
we always get the expected shutdown at the latest when the render call
pings the channel.