Revert "don't overuse requestAnimationFrame (#3657)"
This reverts commit 485c3496a06d9163a4d457efa34369d4f7571f36.

We need to revert because this changes the behavior of JS.show |> JS.show
where the second JS show relied on the first one not being applied yet.

We will keep the change in 1.1 and document it.

This means that https://github.com/phoenixframework/phoenix_live_view/issues/3563 will
be broken again on 1.0.5 :(