recursively undo cloned trees
Fixes: https://github.com/phoenixframework/phoenix_live_view/issues/3684.
Relates to: https://github.com/phoenixframework/phoenix_live_view/pull/3652

The issue was that when we locked a form with an input inside where the
input changed, the form was cloned and then later the input inside was
cloned as well, stored inside the already cloned parent. When undoing,
previously the simple patch copied over the PHX_PRIVATE of the cloned
elements, but with the full DOMPatch, the privates of the source element
were copied, discarding the clone of the input. With this change, we check
if we are undoing locks and manually copy over nested clones to ensure
that they can be correctly applied afterwards.