Ensure `_target` is sent as `"undefined"` when it doesn't exist (#3731)
Fixes #3727.

In LV <= 1.0.5 we sent the `_target` as part of the url encoded form
values, but in 1.0.6 we switched to sending the meta values directly
in the form event to support `JS.push`'s value parameter for nested
values. Since the event itself is JSON encoded, this meant that any keys
with undefined value would be dropped by JS. To be backwards compatible,
we now ensure that the `_target` is always sent and restore the old
behavior of it being the string `"undefined"`, although that seems a bit
ugly. It is what it is now...