use lazy_html for LiveViewTest
Phoenix.LiveView v1.1 will use LazyHTML in its testing infrastructure.
Besides featuring a faster HTML parser, LazyHTML has the benefit of
keeping the HTML document in memory and converting it to a tree lazily,
which allows it to reduce memory allocation (something that can be
better exploited by applications that also choose to migrate to LazyHTML
in their own suites).

Furthermore, it supports additional query selectors and provides a
behaviour more aligned to browsers.
Using this branch in Livebook reduced the test suite from ~9s to ~7s.

Co-Authored-By: Jonatan Kłosko <jonatanklosko@gmail.com>
Co-Authored-By: José Valim <jose.valim@dashbit.co>