Catch promise rejections from pushWithReply (#3766)
* Catch promise rejections from pushWithReply

https://github.com/phoenixframework/phoenix_live_view/pull/3443
changed the pushWithReply function to use promises instead. Previously,
if we did not handle the timeout, it was just ignored, but now we reject
the promise, which will cause unhandled promise rejections.

This PR adds catch clauses to all cases where we previously did not handle
the timeout. At the moment, these log an error, but we can also silently
drop if we want. cc @chrismccord

Supersedes #3760.

* don't catch for pushHookEvent