Fix used_input? after form recovery
There's an inconsistency with using `Phoenix.Component.used_input?`
after form recovery following a socket reconnect.

Suppose we have a field named `"a"` that has been used (i.e. there is no
`"_unused_a"` parameter sent to the server and `used_input?` returns
`true`). What happens is that, immediately after a reconnect, the used
field is initially treated as unused, with an `"_unused_a"` parameter
sent to the server and with `used_input?` incorrectly returning `false`.
However, on the first user interaction with any field in the form, the
correct state is considered and now the filed is again treated as used
(therefore no `"_unused_a"` parameter is sent to the server). Now,
`used_input?` returns `true` as it did before the recovery.

This inconsistent behavior results for example in displayed error
messages suddenly disappearing on socket reconnect, only to reappear on
the first user interaction with the form. This breaks the user
experience and the paradigm which is also mentioned in the documentation
for `used_input?`: "Used inputs are only those inputs that have been
focused, interacted with, or submitted by the client. In most cases, a
user shouldn't receive error feedback for forms they haven't yet
interacted with, until they submit the form. Filtering the errors based
on used input fields can be done with `used_input?/1`."

This patch fixes this issue by immediately copying the DOM privates
(which track, among other things, whether the field has been used) so
that `"_unused_"` params are not sent to the server on socket reconnect
for fields that were considered as used before the disconnect.