fix accumulation of empty text nodes in streams
Closes
https://github.com/phoenixframework/phoenix_live_view/issues/3784.

Because stream nodes are not indexed by morphdom,
new nodes that do not have an ID are just appended to the
stream container. This leads to linear growth of text nodes.

There's a critical problem though as soon as locked trees are involved.
When unlocking, the number of text nodes would double, leading to
exponential growth of (in most cases empty) text nodes.

We already handled this problem for phx-update "append" / "prepend",
but apparently forgot that this also happens in streams.