introduce `:key`-ed `:for` comprehensions
A keyed for comprehension works by transforming the comprehension's
content into a LiveComponent, which will perform change-tracking and
optimizes the payload over the wire:

```heex
<ul>
  <li :for={%{id: @id, name: @name} <- @items} :key={@id}>
    Count: <span>{@count}</span>,
    item: {@name}
  </li>
</ul>
```

To support change tracking, keyed comprehensions require you to define
the left-hand side of the `:for` comprehension, as well as the `:key`,
using assign-syntax: `{@item <- @items}` instead of `{item <- @items}`.

Because they use live components under the hood, keyed comprehensions
come with the limitation of only being supported on HTML tags, so while
you can use a regular `:for` comprehension on `<.function_components>`,
and `<:slots>`, this is not supported keyed comprehensions and will raise
an exception at compile time.