introduce `:key`-ed `:for` comprehensions (#3827)
* introduce `:key`-ed `:for` comprehensions

A keyed for comprehension works by transforming the comprehension's
content into a LiveComponent, which will perform change-tracking and
optimizes the payload over the wire:

```heex
<ul>
  <li :for={%{id: @id, name: @name} <- @items} :key={@id}>
    Count: <span>{@count}</span>,
    item: {@name}
  </li>
</ul>
```

To support change tracking, keyed comprehensions require you to define
the left-hand side of the `:for` comprehension, as well as the `:key`,
using assign-syntax: `{@item <- @items}` instead of `{item <- @items}`.

Because they use live components under the hood, keyed comprehensions
come with the limitation of only being supported on HTML tags, so while
you can use a regular `:for` comprehension on `<.function_components>`,
and `<:slots>`, this is not supported keyed comprehensions and will raise
an exception at compile time.

* add diff test

* add documentation

* introduce :key-ed :for comprehensions without special syntax (#3837)

* with inner_block

* without inner_block

* update docs

* format

* include module and line in key

* store change tracked vars in vars

* Update lib/phoenix_live_view/engine.ex

* Update guides/server/assigns-eex.md

Co-authored-by: José Valim <jose.valim@dashbit.co>

* simplify

* include column in id

* this is nice

* don't classify tag again

* no need for skip_require any more

* ensure meta is a keyword list

* don't duplicate if handling

* remove moved comment

* we don't need render_with

* test for keyed comprehension handling

* unit test

* mark as root in meta

* fix nested change tracking and add test

---------

Co-authored-by: José Valim <jose.valim@dashbit.co>