Ensure change tracking is enabled when traversing keyed comprehensions (#3904)
When rendering live component that shared statics, we'd traverse with change tracking
disabled. The following situation could happen:

LiveComponent 1 renders a keyed comprehension with one entry.
LiveComponent 2 renders a keyed comprehension with more entries, but the first fingerprint
still matches. Now, when rendering the second entry (which is new), we'd build a new
template map, which could be incompatible with the first component. Then, when rendering,
the first entry of the keyed comprehension could point to invalid statics.

The solution is to always consider elements as new when change tracking is disabled.