Fix form recovery when inputs_for is nested in a fieldset
When `inputs_for` is nested in a fieldset, too many copies of the
`sort_param` field are sent to the server upon form recovery, resulting
in additional unwanted dynamic fields. This is the case for example with
the popular Backpex package, when using the `InlineCRUD` field.

This happens since 9a0c375, because of how forms for recovery are cloned
in `view.js`.