Deep clone form for form recovery
The previous algorithm had a flaw:

```javascript
const clonedForm = form.cloneNode(false);
DOM.copyPrivates(clonedForm, form);
Array.from(form.elements).forEach((el) => {
  morphdom(clonedEl, el);
  DOM.copyPrivates(clonedEl, el);
  clonedForm.appendChild(clonedEl);
}
```

While this kept the size of the clonedForm minimal and includes all elements,
it did not consider `<fieldset>` elements. If you had something like this:

```html
<form>
  <fieldset>
    <input name="foo" value="bar">
  </fieldset>
</form>
```

`form.elements` would return `[fieldset, input]`, thus we would actually
include the input twice.

Furthermore, imagine someone did `<fieldset disabled>` which disables
all inputs inside and should exclude them from the form payload. But with
the old code, LiveView would copy the input separately. Because it is not
disabled, it would be included in the recovery payload.

The fix is to actually copy the whole form and then query the DOM for
any associated elements `[form=${form.id}]` to include them.

In the future, we should consider simplifying this code to serialize the
form early and not need to clone DOM nodes, but for now, I think it is fine.

Closes #3914.

Co-Authored-By: Giacomo Mazzamuto <g.mazzamuto@gmail.com>