Consider defaults as changed
Closes https://github.com/phoenixframework/phoenix_live_view/issues/3919.

When a component is called and it receives dynamic attributes

```heex
<.my_component {@action[:attrs] || %{}}>{@action.text}</.my_component>
```

defined as

```elixir
attr :special, :boolean, default: false

defp my_component(assigns) do
  ~H"""
  <div style={if(@special, do: "background-color: red;")}>
    {render_slot(@inner_block)}
  </div>
  """
end
```

Now assuming it is rendered first with `@action` as `%{text: "No red"}`
and then toggling between that and `%{text: "Red", attrs: %{special: true}`:

Previously, once toggled, the background color would be stuck on red,
since `special` would not be considered changed, therefore the diff
for the style would not be sent.

We fix this by setting all defaults to be changed when they are not
explicitly given with the same value as the default.