Ensure component only patches are not lost due to locked tree
Closes https://github.com/phoenixframework/phoenix_live_view/issues/3941.
Possibly related to https://github.com/phoenixframework/phoenix_live_view/issues/3614.

When a component-only diff was applied, we did not account for the
component being part of a locked tree. Thus, when the tree was unlocked,
the component content was unintentionally restored to a previous state.

We solve this by checking for a locked tree and applying the patch
to that instead.