Fix change tracking docs on `:key`
Given the default `:key` is the index of the element, appending should cause only the new element to be sent to the client.

Conversely, if an item is prepended, or, more generally, inserted anywhere within the list, then all subsequent elements will be considered changed. In the worse case that means all elements will be sent to the client.

See https://hexdocs.pm/phoenix_live_view/Phoenix.LiveView.Engine.html#module-comprehensions, which correctly talks about prepeding, not appending.