add attribute tracking (poc)
Adds `phx-update="track-attributes` and `phx-track-attributes` attribute
where the first one tracks attributes on the whole subtree and the second
one only for the element that set the attribute.

Using this makes LiveView keep clientside changes until a server overrides
an attribute value with a new one. This is useful when clientside libraries
rely on attributes they set, but `phx-update="ignore"` is too much.

This is a proof of concept, I'm not sure we actually want to go with
this. For simplicitly, I vendored morphdom into the repo. The important
change there is a new `proxyFromEl` callback + changes to morphAttrs.

Relates to: https://github.com/phoenixframework/phoenix_live_view/issues/4049
Relates to: https://github.com/phoenixframework/phoenix_live_view/pull/4050#issuecomment-3517518544